Pen and Progress bar

+++ VocabularySishu/Model/VSList.h
@@ -41,7 +41,7 @@
 
 + (VSList *)firstList;
 
-- (float)finishProgress;
+- (double)finishProgress;
 
 - (int)rememberedCount;
